DEBUG + TeamIds Update
Fixed Team 0 now being selected in teamSelectScreen by default.

Modified CastlePosGenerator, now it assigns each castle to a random player, instead of in order (previously causing the host to always have a more fixed castle pos)

Added DevTool bool "DEBUG_instaStartMatch" which when true skips the teamSelect screen while in the editor and autoselectects team 0.

Updated CoalitionManager to fill gaps in teams by forcing teams from 0 to playerCount - 1.
(if the teams are are: 1,3,-1,0 it will be converted to 1,2,-1,0 > removed the gap by turning team 3 into 2.