Fixed Some more shit.
Fixed bug in GridManager causing only the server to generate a grid.
Set max value of mapId in GameSettings to 3, because we only have 4 maps.

Fixed Joingame text animation and made text white and bigger.

Gave everyone a Give Turn button, to get the turn when pressed.
Space to revel map works for builds.